+++ spoc/views.py
@@ -73,7 +73,7 @@ def location_list(request):
         return Response(serializer.data)
 
 
-@api_view(['GET'])
+@api_view(['GET', 'POST'])
 def location_detail(request, pk):
     """
     Retrieve a location from LocatonHeader table.
@@ -87,6 +87,12 @@ def location_detail(request, pk):
         serializer = serializers.LocationSerializer(
             location, context={'request': request})
         return Response(serializer.data)
+    elif request.method == 'POST':
+        serializer = serializers.LocationSerializer(location, data=request.DATA)
+        if serializer.is_valid():
+            serializer.save()
+            return Response(serializer.data)
+        return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
 
 
 @api_view(['GET'])
@@ -121,7 +127,7 @@ def scadalocation_detail(request, pk):
         return Response(serializer.data)
 
 
-@api_view(['GET'])
+@api_view(['GET', 'POST'])
 def header_detail(request, pk):
     """
     Retrieve a header details from Header table.
@@ -135,140 +141,9 @@ def header_detail(request, pk):
         serializer = serializers.HeaderSerializer(
             header, context={'request': request})
         return Response(serializer.data)
-
-
-@api_view(['GET'])
-def oei_list(request):
-    """
-    List all locations.
-    """
-    if request.method == 'GET':
-        locations = models.OEI.objects.all()
-        serializer = serializers.OEISerializer(locations, many=True,
-                                               context={'request': request})
-        return Response(serializer.data)
-
-
-@api_view(['GET'])
-def oei_detail(request, pk):
-    """
-    Retrieve a location from oei table.
-    """
-    try:
-        location = models.OEI.objects.get(pk=pk)
-    except models.OEI.DoesNotExist:
-        return Response(status=status.HTTP_404_NOT_FOUND)
-
-    if request.method == 'GET':
-        serializer = serializers.OEISerializer(location, context={'request': request})
-        return Response(serializer.data)
-
-
-@api_view(['GET'])
-def wnsattribute_list(request):
-    """
-    List all wms-attributes.
-    """
-    if request.method == 'GET':
-        wnsattributes = models.WNSAttribute.objects.all()
-        serializer = serializers.WNSAttributeSerializer(wnsattributes, many=True,
-                                                        context={'request': request})
-        return Response(serializer.data)
-
-
-@api_view(['GET'])
-def wnsattribute_detail(request, pk):
-    """
-    Retrieve a wns-attribute.
-    """
-    try:
-        wnsattribute = models.WNSAttribute.objects.get(pk=pk)
-    except models.WNSAttribute.DoesNotExist:
-        return Response(status=status.HTTP_404_NOT_FOUND)
-
-    if request.method == 'GET':
-        serializer = serializers.WNSAttributeSerializer(wnsattribute,
-                                                        context={'request': request})
-        return Response(serializer.data)
-
-
-# @api_view(['GET', 'POST'])
-# def location_list(request):
-#     """
-#     List all locations, or create a new location.
-#     """
-#     if request.method == 'GET':
-#         locations = models.Location.objects.all()
-#         serializer = serializers.LocationSerializer(locations, many=True,
-#                                                     context={'request': request})
-#         return Response(serializer.data)
-
-#     elif request.method == 'POST':
-#         data = JSONParser().parse(request)
-#         serializer = serializers.LocationSerializer(data=data)
-#         if serializer.is_valid():
-#             serializer.save()
-#             return Response(serializer.data, status=status.HTTP_201_CREATED)
-#         return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
-
-
-# @api_view(['GET'])
-# def location_detail(request, pk):
-#     """
-#     Retrieve, update or delete a location.
-#     """
-#     try:
-#         location = models.Location.objects.get(pk=pk)
-#     except models.Location.DoesNotExist:
-#         return Response(status=status.HTTP_404_NOT_FOUND)
-    
-#     if request.method == 'GET':
-#         serializer = serializers.LocationSerializer(location, context={'request': request})
-#         return Response(serializer.data)
-
-        
-# @api_view(['GET', 'PUT', 'DELETE'])
-# def location_detail(request, pk):
-#     """
-#     Retrieve, update or delete a location.
-#     """
-#     try:
-#         location = models.Location.objects.get(pk=pk)
-#     except models.Location.DoesNotExist:
-#         return Response(status=status.HTTP_404_NOT_FOUND)
-
-#     if request.method == 'GET':
-#         serializer = serializers.LocationSerializer(location)
-#         return Response(serializer.data)
-
-#     elif request.method == 'PUT':
-#         serializer = serializers.LocationSerializer(location, data=request.DATA)
-#         if serializer.is_valid():
-#             serializer.save()
-#             return Response(serializer.data)
-#         return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
-
-#     elif request.method == 'DELETE':
-#         location.delete()
-#         return HttpResponse(status=status.HTTP_204_NO_CONTENT)
-
-
-class ListLocations(viewsets.ModelViewSet):
-    """ List dummy locations. """
-
-    model = models.Location
-    permission_classes = (permissions.IsAuthenticatedOrReadOnly,)
-
-
-class LocationDetails(viewsets.ModelViewSet):
-    """ List dummy locations. """
-
-    model = models.Location
-    permission_classes = (permissions.IsAuthenticatedOrReadOnly,)
-
-    def get(self, request, pk=None):
-        location = models.Locations.objects.get(pk=pk)
-        return Response(location)
-
-    def post(self, request, pk=None, *args, **kwargs):
-        return Response(models.Location.objects.get(pk=pk))
+    elif request.method == 'POST':
+        serializer = serializers.HeaderSerializer(header, request.DATA)
+        if serializer.is_valid():
+            serializer.save()
+            return Response(serializer.data)
+        return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)